meal-prep-website/backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
sqlalchemy>=2.0.0
pydantic>=2.4.2
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
anthropic>=0.5.0
python-multipart>=0.0.6
